OCPBUGS-44576: Resolve cycles in actions
