+++ frontend/public/components/app.jsx
@@ -12,7 +12,8 @@ import 'abort-controller/polyfill';
 import store, { applyReduxExtensions } from '../redux';
 import { useTranslation } from 'react-i18next';
 import { coFetchJSON, appInternalFetch } from '../co-fetch';
-import { detectFeatures, setFlag } from '../actions/features';
+import { detectFeatures } from '../actions/features';
+import { setFlag } from '../actions/flagActions';
 import AppContents from './app-contents';
 import { Masthead } from './masthead';
 import { getBrandingDetails } from './masthead-branding';
