+++ frontend/public/reducers/features.ts
@@ -28,7 +28,7 @@ import { referenceForGroupVersionKind, referenceForModel } from '../module/k8s/k
 import { RootState } from '../redux';
 import { ActionType as K8sActionType } from '@console/dynamic-plugin-sdk/src/app/k8s/actions/k8s';
 import { FeatureState } from '@console/dynamic-plugin-sdk/src/app/features';
-import { FeatureAction, ActionType } from '../actions/features';
+import { FeatureAction, ActionType } from '../actions/flagActions';
 import { pluginStore } from '../plugins';
 
 // eslint-disable-next-line prettier/prettier
